feat: agregar componentes de CombatStats, EncyclopediaImage, ItemHeader, y ItemProperties
